add end to end tests for time restrictions
