+++ repos/fdbt-site/cypress_tests/cypress/integration/globalSettings/timeRestrictions.ts
@@ -1,0 +1,17 @@
+import { deleteAllCards, startGlobalSettings } from '../../support/globalSettings';
+import { createEditTimeRestriction } from '../../support/timeRestrictions';
+
+describe('time restrictions', () => {
+    it('creates edits and deletes time restrictions', () => {
+        startGlobalSettings();
+
+        cy.contains('Time restrictions').click();
+
+        // start with clean environment
+        deleteAllCards();
+
+        createEditTimeRestriction();
+
+        deleteAllCards();
+    });
+});
