+++ repos/fdbt-site/cypress_tests/cypress/support/steps.ts
@@ -18,9 +18,7 @@ import {
     completeMultipleProducts,
     completeOperatorSearch,
     clickSomeCheckboxes,
-    randomlyChooseASchoolProof,
     randomlyDecideTermRestrictions,
-    randomlyChooseSchoolAgeLimits,
     randomlyChooseProductPeriodValidity,
 } from './helpers';
