+++ repos/fdbt-site/src/pages/viewTimeRestrictions.tsx
@@ -62,7 +62,7 @@ const ViewTimeRestrictions = ({ timeRestrictions, referer, csrfToken }: TimeRest
     return (
         <BaseLayout title={title} description={description} showNavigation referer={referer}>
             <div className="govuk-width-container">
-                <main className="govuk-main-wrapper">
+                <main className="govuk-main-wrapper" data-card-count={timeRestrictions.length}>
                     <div className="govuk-grid-row">
                         <div className="govuk-grid-column-one-third">
                             <SubNavigation />
